small fix to get rid of gcc warning: 'curl_easy_setopt expects a long argument [-Wattribute-warning]'

+++ lib/src/gretl_www.c
@@ -527,10 +527,10 @@ static int common_curl_setup (CURL **pcurl)
         err = 1;
     } else {
 #if WDEBUG
-        curl_easy_setopt(*pcurl, CURLOPT_VERBOSE, 1);
+        curl_easy_setopt(*pcurl, CURLOPT_VERBOSE, (long int) 1);
 #else
         curl_easy_setopt(*pcurl, CURLOPT_VERBOSE,
-                         getenv("GRETL_WWW_VERBOSE") != NULL);
+                         (long int) (getenv("GRETL_WWW_VERBOSE") != NULL));
 #endif
 #ifdef _WIN32
         /* be on the safe side: 'http' can turn into 'https'
